+++ src/components/draft/DraftCard.js
@@ -1,7 +1,16 @@
 import * as S from "styles/DraftCard.styles";
 
-export const DraftCard = ({ data, sport }) => {
-  const { round, pick, overallPick, player, position, team } = data;
+export const DraftCard = ({ data, sport, isCompleted }) => {
+  const {
+    round,
+    pick,
+    overallPick,
+    player = "",
+    position,
+    team,
+    tradedTo = false,
+    fantasyTeam,
+  } = data;
   const playerNameArray = player.split(" ");
   const playerName =
     playerNameArray.length > 1
@@ -12,13 +21,25 @@ export const DraftCard = ({ data, sport }) => {
     <S.PaddingContainer sport={sport} position={position}>
       <S.DraftPickNumberText>{`${round}-${pick} (${overallPick})`}</S.DraftPickNumberText>
 
-      <S.ExtraContainer>
-        <S.TooltipContainer className="hover-text">
-          <S.PlayerText>{`${playerName}`}</S.PlayerText>
-          <S.TooltipText className="tooltip-text">{player}</S.TooltipText>
-        </S.TooltipContainer>
-      </S.ExtraContainer>
-      <S.DraftPickNumberText>{`${position} - ${team}`}</S.DraftPickNumberText>
+      {isCompleted && (
+        <>
+          <S.ExtraContainer>
+            <S.TooltipContainer className="hover-text">
+              <S.PlayerText>{`${playerName}`}</S.PlayerText>
+              <S.TooltipText className="tooltip-text">{player}</S.TooltipText>
+            </S.TooltipContainer>
+          </S.ExtraContainer>
+          <S.DraftPickNumberText>{`${position} - ${team}`}</S.DraftPickNumberText>
+        </>
+      )}
+      {tradedTo && (
+        <>
+          <S.ExtraContainer>
+            <S.PlayerText>{`to ${tradedTo}`}</S.PlayerText>
+          </S.ExtraContainer>
+          <S.DraftPickNumberText>{`via ${fantasyTeam}`}</S.DraftPickNumberText>
+        </>
+      )}
     </S.PaddingContainer>
   );
 };